requirements.txt
django
django-activelink
django-bootstrap-form
django-extensions
django-filter
django-taggit
